Bellabeat Fitness
I used R programming to gain insights on fitness data from Kaggle. I installed the tidyverse package and used the head a colnames functions to get a preview of the datasets. I used pipes to get a summary of the datasets and the ggplot function to create visualizations and see any trends between the datasets